moving atom_data
