+++ docs/physics/plasma/plasma_plots/nebular_ionization_balance.py
@@ -3,8 +3,7 @@ import os
 from matplotlib import colors
 
 import tardis.util.base
-from tardis import util
-from tardis.io import atomic
+from tardis.io.atom_data import base
 from matplotlib import pyplot as plt
 
 import numpy as np
@@ -22,7 +21,7 @@ ax2 = plt.figure(2).add_subplot(111)
 atom_fname = os.path.expanduser('~/.tardis/si_kurucz.h5')
 
 # reading in the HDF5 File
-atom_data = atomic.AtomData.from_hdf(atom_fname)
+atom_data = base.AtomData.from_hdf(atom_fname)
 
 #The atom_data needs to be prepared to create indices. The Class needs to know which atomic numbers are needed for the
 #calculation and what line interaction is needed (for "downbranch" and "macroatom" the code creates special tables)
