+++ tardis/io/atom_data/__init__.py
@@ -1,0 +1,1 @@
+from tardis.io.atom_data.base import AtomData
